feat: Add support for parsing S3 access logs (#33)
This adds support for processing S3 Server Access Logs written with the
date-based partitioning scheme. Because these are written in
non-compressed format this also adds logic to determine if a log object
is gzip-compressed or not. So this _should_ address #31 but I don't have
a way to test this in my stack.